If the file already exists in destination, do not overwrite file

+++ src/main/java/com/example/fileorganizer/fileorganizer/ScheduledTask.java
@@ -7,10 +7,10 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import java.io.File;
-import java.io.IOException;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
@@ -18,8 +18,8 @@ import java.util.Date;
 public class ScheduledTask {
     private static final Logger log = LoggerFactory.getLogger(ScheduledTask.class);
     private static final SimpleDateFormat simpleDateFormat = new SimpleDateFormat("HH:mm:ss");
-
-//    touch file.txt file.mp3 file.mp4 file.jpg file.png file.rtf file.wav file.mov
+    @Value("${DOWNLOADS_DIRECTORY}")
+    private String downloadsDirectory;
 
     @Value("${PICTURES_DIRECTORY}")
     private String picturesDirectory;
@@ -30,9 +30,6 @@ public class ScheduledTask {
     @Value("${DOCUMENTS_DIRECTORY}")
     private String documentsDirectory;
 
-    @Value("${DOWNLOADS_DIRECTORY}")
-    private String downloadsDirectory;
-
     @Value("${MUSIC_DIRECTORY}")
     private String musicDirectory;
 
@@ -40,9 +37,12 @@ public class ScheduledTask {
     public void reportCurrentTime() {
         log.info("The time is now {}", simpleDateFormat.format(new Date()));
 
-        // Check Downloads directory for files
         log.info("Downloads directory: {}", downloadsDirectory);
         log.info("Music directory: {}", musicDirectory);
+        log.info("Documents directory: {} ", documentsDirectory);
+        log.info("Pictures directory: {} ", picturesDirectory);
+        log.info("Videos directory: {}", videosDirectory);
+
         checkDirectories();
     }
 
@@ -50,10 +50,12 @@ public class ScheduledTask {
         Path path = FileSystems.getDefault().getPath(downloadsDirectory);
 
         if (Files.isDirectory(path)) {
-            log.info("Directory exists");
+            log.info("The {} directory exists", downloadsDirectory);
             checkFiles();
         } else {
-            log.info("Directory {} does not exist", downloadsDirectory);
+            log.info("The {} directory does not exist", downloadsDirectory);
+            File newDirectory = new File(downloadsDirectory);
+            newDirectory.mkdir();
         }
     }
 
@@ -61,26 +63,46 @@ public class ScheduledTask {
         File[] files = new File(downloadsDirectory).listFiles();
 
         for (File file : files) {
-            log.info("file name: {}", file.getName());
+            log.info("Filename: {}", file.getName());
 
             if (file.getName().endsWith(".mp3") || file.getName().endsWith(".wav")) {
-                log.info("music file, moved to the music directory: {}", musicDirectory);
-                file.renameTo(new File(musicDirectory + File.separator + file.getName()));
-            }
-
-            if (file.getName().endsWith(".jpg") || file.getName().endsWith(".png")) {
-                log.info("image file, moved to the music directory: {}", picturesDirectory);
-                file.renameTo(new File(picturesDirectory + File.separator + file.getName()));
-            }
-
-            if (file.getName().endsWith(".rtf") || file.getName().endsWith(".txt")) {
-                log.info("text file, moved to the music directory: {}", documentsDirectory);
-                file.renameTo(new File(documentsDirectory + File.separator + file.getName()));
+                if (Files.exists(Paths.get(musicDirectory + File.separator + file.getName()))) {
+                    log.info("The file {} exists in the destination directory {}", file.getName(), musicDirectory);
+                } else {
+                    log.info("The file {} does not exist in the destination directory {}", file.getName(), musicDirectory);
+                    file.renameTo(new File(musicDirectory + File.separator + file.getName()));
+                    log.info("The file {} was moved to the directory: {}", file.getName(), musicDirectory);
+                }
             }
 
             if (file.getName().endsWith(".mov") || file.getName().endsWith(".mp4")) {
-                log.info("videos file, moved to the music directory: {}", videosDirectory);
-                file.renameTo(new File(videosDirectory + File.separator + file.getName()));
+                if (Files.exists(Paths.get(videosDirectory + File.separator + file.getName()))) {
+                    log.info("The file {} exists in the destination directory {}", file.getName(), videosDirectory);
+                } else {
+                    log.info("The file {} does not exist in the destination directory {}", file.getName(), videosDirectory);
+                    file.renameTo(new File(videosDirectory + File.separator + file.getName()));
+                    log.info("The file {} was moved to the directory: {}", file.getName(), videosDirectory);
+                }
+            }
+
+            if (file.getName().endsWith(".jpg") || file.getName().endsWith(".png")) {
+                if (Files.exists(Paths.get(picturesDirectory + File.separator + file.getName()))) {
+                    log.info("The file {} exists in the destination directory {}", file.getName(), picturesDirectory);
+                } else {
+                    log.info("The file {} does not exist in the destination directory {}", file.getName(), picturesDirectory);
+                    file.renameTo(new File(picturesDirectory + File.separator + file.getName()));
+                    log.info("The file {} was moved to the directory: {}", file.getName(), picturesDirectory);
+                }
+            }
+
+            if (file.getName().endsWith(".rtf") || file.getName().endsWith(".txt")) {
+                if (Files.exists(Paths.get(documentsDirectory + File.separator + file.getName()))) {
+                    log.info("The file {} exists in the destination directory {}", file.getName(), documentsDirectory);
+                } else {
+                    log.info("The file {} does not exist in the destination directory {}", file.getName(), documentsDirectory);
+                    file.renameTo(new File(documentsDirectory + File.separator + file.getName()));
+                    log.info("The file {} was moved to the directory: {}", file.getName(), documentsDirectory);
+                }
             }
         }
     }
